Create Lonely Integer - Bash!
There are  integers in an array . All but one integer occur in pairs. Your task is to find the number that occurs only once.

Input Format

The first line of the input contains an integer , indicating the number of integers. The next line contains  space-separated integers that form the array .